requirements.txt
aiofiles==22.1.0
aiohttp==3.8.4
aioshutil==1.3
aiosignal==1.3.1
anyio==3.6.2
apify==1.1.1
apify-client==1.3.0
async-timeout==4.0.2
attrs==23.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
cryptography==39.0.2
dataclasses-json==0.5.7
dnspython==2.3.0
frozenlist==1.3.3
h11==0.14.0
httpcore==0.16.3
httpx==0.24.1
idna==3.4
langchain==0.1.9
langchain-community==0.0.34
langchain-openai==0.0.8
loguru>=0.7.2,<0.8.0
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.3
openai==1.12.0
openapi-schema-pydantic==1.2.4
packaging>=23.2,<24.0
pinecone-client==3.0.0
psutil==5.9.5
pycparser==2.21
pydantic==1.10.7
pyee==9.0.4
python-dateutil==2.8.2
PyYAML==6.0
regex==2023.5.5
requests>=2.31,<3.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
sortedcollections==2.1.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.12
tenacity==8.2.2
tiktoken
fastembed
tqdm==4.65.0
typing-inspect==0.8.0
typing_extensions>=4.7,<5
urllib3==1.26.15
websockets==10.4
yarl==1.9.2